+++ load.zsh
@@ -26,12 +26,52 @@ add-zsh-hook preexec slick_prompt_preexec
 # Register zle widgets
 zle -N zle-keymap-select
 zle -N zle-line-init
+zle -N accept-line slick_prompt_accept_line
 
 # Global variables
 typeset -g slick_prompt_data
 typeset -g slick_prompt_fd
 typeset -g slick_prompt_timestamp
 typeset -g slick_prompt_elapsed
+
+function slick_prompt_transient_enabled {
+    [[ "${SLICK_PROMPT_TRANSIENT:-1}" != "0" ]]
+}
+
+function slick_prompt_rfc3339_timestamp {
+    local timestamp
+
+    strftime -s timestamp '%Y-%m-%dT%H:%M:%S%z' $EPOCHSECONDS
+    print -r -- "${timestamp[1,-3]}:${timestamp[-2,-1]}"
+}
+
+function slick_prompt_render {
+    local exit_status=${1:-0}
+    local transient=${2:-0}
+    local transient_timestamp=${3:-}
+    local -a args
+
+    args=(
+        "$SLICK_PATH"
+        prompt
+        -k "${KEYMAP:-main}"
+        -r "$exit_status"
+        -d "${slick_prompt_data:-}"
+    )
+
+    if [[ -n "${slick_prompt_elapsed:-}" ]]; then
+        args+=(-e "$slick_prompt_elapsed")
+    fi
+
+    if [[ "$transient" == 1 ]]; then
+        args+=(--transient)
+        if [[ -n "$transient_timestamp" ]]; then
+            args+=(--transient-timestamp "$transient_timestamp")
+        fi
+    fi
+
+    "${args[@]}"
+}
 
 function slick_prompt_refresh {
     local exit_status=$?
@@ -41,15 +81,7 @@ function slick_prompt_refresh {
     # ZSH will call this function again if there's more data
     if read -r -u $1 line; then
         slick_prompt_data="$line"
-
-        # Always pass elapsed time if available (needed for ALL phases to show consistent elapsed time!)
-        # Use the pre-calculated elapsed time from precmd to avoid flickering
-        if [[ -n "$slick_prompt_elapsed" ]]; then
-            PROMPT=$($SLICK_PATH prompt -k "$KEYMAP" -r $exit_status -d ${slick_prompt_data:-""} -e $slick_prompt_elapsed)
-        else
-            PROMPT=$($SLICK_PATH prompt -k "$KEYMAP" -r $exit_status -d ${slick_prompt_data:-""})
-        fi
-
+        PROMPT=$(slick_prompt_render "$exit_status")
         zle && zle reset-prompt
         return  # RETURN immediately - don't block! Handler will be called again for next line
     fi
@@ -60,7 +92,7 @@ function slick_prompt_refresh {
     unset slick_prompt_elapsed
     zle -F $1
     exec {1}<&-
-    
+
     # Reset global fd if it matches
     if [[ "$1" == "$slick_prompt_fd" ]]; then
         unset slick_prompt_fd
@@ -68,8 +100,21 @@ function slick_prompt_refresh {
 }
 
 function zle-line-init zle-keymap-select {
-    PROMPT=$($SLICK_PATH prompt -k "$KEYMAP" -d ${slick_prompt_data:-""})
+    PROMPT=$(slick_prompt_render 0)
     zle && zle reset-prompt
+}
+
+function slick_prompt_accept_line {
+    local exit_status=$?
+    local transient_timestamp
+
+    if slick_prompt_transient_enabled; then
+        transient_timestamp=$(slick_prompt_rfc3339_timestamp)
+        PROMPT=$(slick_prompt_render "$exit_status" 1 "$transient_timestamp")
+        zle reset-prompt
+    fi
+
+    zle .accept-line
 }
 
 function slick_prompt_precmd() {
@@ -86,7 +131,7 @@ function slick_prompt_precmd() {
     # If timestamp is set (command was run), calculate elapsed seconds
     # Otherwise, leave it unset (no command was run, e.g., just pressed enter)
     if [[ -n "$slick_prompt_timestamp" ]]; then
-        slick_prompt_elapsed=$(( $EPOCHSECONDS - $slick_prompt_timestamp ))
+        slick_prompt_elapsed=$(( EPOCHSECONDS - slick_prompt_timestamp ))
         # Ensure elapsed time is never negative (can happen with clock adjustments)
         [[ $slick_prompt_elapsed -lt 0 ]] && slick_prompt_elapsed=0
     else
@@ -109,4 +154,4 @@ function slick_prompt_preexec() {
     echo -ne "\e[4 q"
 }
 
-echo "slick prompt loaded (2-phase: instant [user path branch] + async [git status])"
+echo "slick prompt loaded (2-phase async prompt + transient scrollback by default; set SLICK_PROMPT_TRANSIENT=0 to disable)"